bug: fix incoming session check
I'm reverting a recent "fix" that was inverting the local-remote session
id check - the error in that commit was to change the check to match the
description in the comment. In reality, it was the comment that was
wrong. The reason why we cannot check remote session id in all packets
is because remote-session-id only present if the packet has an ack
array. The handshake was broken by the previous change, but I did not
catch it because the integration tests were not running for the
refactored codebase.

Also, this commit enables integration tests for the new internal path to
avoid this kind of errors in the future.

Some test code was removed, because the script to parse certs etc was
written before minivpn could parse inline credentials. not needed
anymore.